src/crypto-kafka/requirements.txt
dacite
six
kafka-python
kafka-python-ng
prometheus_client
pycoingecko
